Fix ignoreCase flag in src/RegEx/init.lua
## Summary
The RegEx code refers to `flag.ignoreCase` to accomplish ignored-case pattern matching, but the string in the `flag_map` (which is used to set the flag object, see [here](https://github.com/Roblox/luau-regexp/blob/main/src/RegEx/init.lua#L1969)) is labelled 'caseless' instead of 'ignoreCase'.

## Testing:
Tested with a pattern that works for lowercase text but not the identical (but uppercased) text.

Before the change, it fails to match uppercase version. After the change, it matches.